refactor(aggregator): reorganize `gcp_uploader` module
* promoted to a directory
* splited into several part in order to isolate the old backend since it
  will be removed soon
