+++ mithril-aggregator/src/file_uploaders/mod.rs
@@ -4,7 +4,7 @@ mod interface;
 mod local_uploader;
 
 pub use dumb_uploader::*;
-pub use gcp_uploader::{CloudRemotePath, GcpBackendUploader, GcpUploader};
+pub use gcp_uploader::{CloudRemotePath, GcpBackendUploaderLegacy, GcpUploader};
 pub use interface::{FileUploadRetryPolicy, FileUploader};
 pub use local_uploader::LocalUploader;
